+++ refarch-apigateway/apigateway/src/test/java/de/muenchen/refarch/TestConstants.java
@@ -2,7 +2,7 @@
  * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
  * der Landeshauptstadt München, 2024
  */
-package refarch;
+package de.muenchen.refarch;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
