+++ refarch-apigateway/apigateway/src/test/java/de/muenchen/refarch/controller/ActuatorInfoEndpointTest.java
@@ -2,19 +2,18 @@
  * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
  * der Landeshauptstadt München, 2024
  */
-package refarch.filter;
+package de.muenchen.refarch.controller;
 
-import refarch.ApiGatewayApplication;
+import de.muenchen.refarch.ApiGatewayApplication;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
-import static refarch.TestConstants.SPRING_TEST_PROFILE;
+import static de.muenchen.refarch.TestConstants.SPRING_TEST_PROFILE;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(
@@ -22,17 +21,19 @@ import static refarch.TestConstants.SPRING_TEST_PROFILE;
         webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
 )
 @ActiveProfiles(SPRING_TEST_PROFILE)
-class CsrfTokenAppendingHelperFilterTest {
+public class ActuatorInfoEndpointTest {
 
     @Autowired
     private WebTestClient webTestClient;
 
     @Test
-    @WithMockUser
-    void csrfCookieAppendition() {
-        webTestClient.get().uri("/").exchange()
-                .expectHeader()
-                .valueMatches("set-cookie", "XSRF-TOKEN=[a-f\\d]{8}(-[a-f\\d]{4}){3}-[a-f\\d]{12}?;\\sPath=/");
+    public void actuatorInfoProvidesAppswitcherUrl() {
+        //@formatter:off
+        webTestClient.get().uri("/actuator/info").exchange()
+                .expectStatus().isOk()
+                .expectBody()
+                    .jsonPath("$.appswitcher.url").isEqualTo("https://test-url-appswitcher.muenchen.de");
+        //@formatter:on
     }
 
 }
